pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-resume"
version = "0.1.0"
description = "Stateful CLI scheduler for coding agents with persistent prompt queues."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "agent-resume contributors" }]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "PyYAML>=6.0.0",
]

[project.scripts]
agent-resume = "agent_resume.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
